add healthy user test
