+++ src/test/java/user/UserTest.java
@@ -1,10 +1,14 @@
 package user;
 
 import enums.Routine;
+import nutritionalCondition.Diabetic;
+import nutritionalCondition.Hypertensive;
+import nutritionalCondition.NutritionalCondition;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import javax.swing.event.HyperlinkEvent;
 import java.time.LocalDate;
 import java.util.HashSet;
 
@@ -14,6 +18,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 @DisplayName("Given a user")
 public class UserTest {
     User user;
+    NutritionalCondition diabetic = Diabetic.getInstance();
+    NutritionalCondition hypertensive = Hypertensive.getInstance();
 
     @Test
     @DisplayName("its bmi is equal to the user's weight divided by it's squared height")
@@ -24,6 +30,16 @@ public class UserTest {
     @Test
     @DisplayName("is healthy if he has no preexisting nutritional conditions and has a healthy bmi")
     public void healthyUserWithNoPreexistingNutritionalConditions() {
+        assertTrue(user.healthy());
+    }
+
+    @Test
+    @DisplayName("is healthy if all its nutritional conditions are corrected and has a healthy bmi")
+    public void healthyUserCorrectedConditionsHealthyBmi(){
+        user.addNutritionalCondition(diabetic);
+        user.addNutritionalCondition(hypertensive);
+        user.setRoutine(Routine.INTENSE);
+        user.setWeight(70);
         assertTrue(user.healthy());
     }
